Banners and splash screen updated to 2015 (splash now as higher-res JPG, removed load update) New planetary surface manager: - supports resolution level 19 - supports elevation modifier data from Elev_mod subdir - updated method for rendering water surfaces - adjusted hightmap interpolation for patch mesh generation to match vessel elevation Base structures: - mostly fixed to local elevation (some, plus shadows for all still to do) added tileedit tool to utils directory (needs Matlab 2012a runtimes)

+++ Orbitersdk/samples/Atlantis/Atlantis/Atlantis.cpp
@@ -637,7 +637,7 @@ void Atlantis::DefineAnimations (void)
 
 	// ***** 9. SSME pitch gimbal animations
 	double init_gimbal = -10*RAD;
-	double max_gimbal = -0.2*PI;
+	float max_gimbal = (float)(-0.2*PI);
 	anim_ssme = CreateAnimation (init_gimbal/max_gimbal);
 
 	static UINT SSMEL_Grp = GRP_SSMEL;
